Add files via upload
Finished the full 45 seconds of animation with 1 problem to still fix. Added a few sound clips for certain sections (fairy dust sound for protons, bouncing ball sound for the oxygen, and lightsaber turning on for the bonds forming). 

Still need to add more audio clips and fix the hydrogen molecule moving as one object (instead of the hydrogen lagging behind because it followed a nurbs path).